[pal] Improvements
- Improve memory safety and clarify ownership
- Try harder to merge connected features, to better handle situations
like T intersections or roads which divide into two lanes
- Fix labels are sometimes placed in conflict with line features when
the conflicting line has the same label yet we couldn't successfully
merge the two line parts

+++ src/core/pal/pal.h
@@ -264,8 +264,8 @@ namespace pal
       int mTenure = 10;
       double mCandListSize = 0.2;
 
-      long long mNextCandidateId = 1;
-      mutable QHash< QPair< long long, long long >, bool > mCandidateConflicts;
+      unsigned int mNextCandidateId = 1;
+      mutable QHash< QPair< unsigned int, unsigned int >, bool > mCandidateConflicts;
 
       /**
        * \brief show partial labels (cut-off by the map canvas) or not
